Rename button folder
