+++ src/components/product-card.component/product-card.component.jsx
@@ -1,5 +1,5 @@
 import './product-card.styles.scss'
-import Button from '../button.component/button.component'
+import Button from '../button/button.component'
 import { useContext } from 'react'
 import { CartContext } from '../../contexts/cart.context'
