+++ src/components/sign-up-form/sign-up-form.component.jsx
@@ -2,7 +2,7 @@ import { useState } from "react"
 import { createAuthUserWithEmailAndPassword, createUserDocumentFromAuth } from "../../util/firebase/firebase.utils"
 import FormInput from "../form-input/form-input.component"
 import './sign-up-form.styles.scss'
-import Button from "../button.component/button.component"
+import Button from "../button/button.component"
 
 const defaultFormFields = {
     displayName: '',
